# Change Mark to be readonly struct