--- a/YamlDotNet/Core/Events/SequenceEnd.cs
+++ b/YamlDotNet/Core/Events/SequenceEnd.cs
@@ -43,7 +43,7 @@ namespace YamlDotNet.Core.Events
         /// </summary>
         /// <param name="start">The start position of the event.</param>
         /// <param name="end">The end position of the event.</param>
-        public SequenceEnd(Mark start, Mark end)
+        public SequenceEnd(in Mark start, in Mark end)
             : base(start, end)
         {
         }