--- a/YamlDotNet/Core/Events/StreamStart.cs
+++ b/YamlDotNet/Core/Events/StreamStart.cs
@@ -63,7 +63,7 @@ namespace YamlDotNet.Core.Events
         /// </summary>
         /// <param name="start">The start position of the event.</param>
         /// <param name="end">The end position of the event.</param>
-        public StreamStart(Mark start, Mark end)
+        public StreamStart(in Mark start, in Mark end)
             : base(start, end)
         {
         }