--- a/YamlDotNet/Core/ForwardAnchorNotSupportedException.cs
+++ b/YamlDotNet/Core/ForwardAnchorNotSupportedException.cs
@@ -41,7 +41,7 @@ namespace YamlDotNet.Core
         /// <summary>
         /// Initializes a new instance of the <see cref="AnchorNotFoundException"/> class.
         /// </summary>
-        public ForwardAnchorNotSupportedException(Mark start, Mark end, string message)
+        public ForwardAnchorNotSupportedException(in Mark start, in Mark end, string message)
             : base(start, end, message)
         {
         }