--- a/YamlDotNet/Core/Scanner.cs
+++ b/YamlDotNet/Core/Scanner.cs
@@ -2207,7 +2207,7 @@ namespace YamlDotNet.Core
         ///      %TAG    !yaml!  tag:yaml.org,2002:  \n
         ///       ^^^
         /// </summary>
-        private string ScanDirectiveName(Mark start)
+        private string ScanDirectiveName(in Mark start)
         {
             var name = new StringBuilder();
 
@@ -2252,7 +2252,7 @@ namespace YamlDotNet.Core
         ///      %YAML   1.1     # a comment \n
         ///           ^^^^^^
         /// </summary>
-        private Token ScanVersionDirectiveValue(Mark start)
+        private Token ScanVersionDirectiveValue(in Mark start)
         {
             SkipWhitespaces();
 
@@ -2283,7 +2283,7 @@ namespace YamlDotNet.Core
         ///      %TAG    !yaml!  tag:yaml.org,2002:  \n
         ///          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
         /// </summary>
-        private Token ScanTagDirectiveValue(Mark start)
+        private Token ScanTagDirectiveValue(in Mark start)
         {
             SkipWhitespaces();
 
@@ -2371,7 +2371,7 @@ namespace YamlDotNet.Core
         /// Decode an URI-escape sequence corresponding to a single UTF-8 character.
         /// </summary>
 
-        private string ScanUriEscapes(Mark start)
+        private string ScanUriEscapes(in Mark start)
         {
             // Decode the required number of characters.
 
@@ -2495,7 +2495,7 @@ namespace YamlDotNet.Core
         ///      %YAML   1.1     # a comment \n
         ///                ^
         /// </summary>
-        private int ScanVersionDirectiveNumber(Mark start)
+        private int ScanVersionDirectiveNumber(in Mark start)
         {
             var value = 0;
             var length = 0;