--- a/YamlDotNet/Core/Tokens/BlockEntry.cs
+++ b/YamlDotNet/Core/Tokens/BlockEntry.cs
@@ -39,7 +39,7 @@ namespace YamlDotNet.Core.Tokens
         /// </summary>
         /// <param name="start">The start position of the token.</param>
         /// <param name="end">The end position of the token.</param>
-        public BlockEntry(Mark start, Mark end)
+        public BlockEntry(in Mark start, in Mark end)
             : base(start, end)
         {
         }