--- a/YamlDotNet/Core/Tokens/BlockMappingStart.cs
+++ b/YamlDotNet/Core/Tokens/BlockMappingStart.cs
@@ -39,7 +39,7 @@ namespace YamlDotNet.Core.Tokens
         /// </summary>
         /// <param name="start">The start position of the token.</param>
         /// <param name="end">The end position of the token.</param>
-        public BlockMappingStart(Mark start, Mark end)
+        public BlockMappingStart(in Mark start, in Mark end)
             : base(start, end)
         {
         }