--- a/YamlDotNet/Core/Tokens/DocumentEnd.cs
+++ b/YamlDotNet/Core/Tokens/DocumentEnd.cs
@@ -39,7 +39,7 @@ namespace YamlDotNet.Core.Tokens
         /// </summary>
         /// <param name="start">The start position of the token.</param>
         /// <param name="end">The end position of the token.</param>
-        public DocumentEnd(Mark start, Mark end)
+        public DocumentEnd(in Mark start, in Mark end)
             : base(start, end)
         {
         }