--- a/YamlDotNet/Core/Tokens/DocumentStart.cs
+++ b/YamlDotNet/Core/Tokens/DocumentStart.cs
@@ -39,7 +39,7 @@ namespace YamlDotNet.Core.Tokens
         /// </summary>
         /// <param name="start">The start position of the token.</param>
         /// <param name="end">The end position of the token.</param>
-        public DocumentStart(Mark start, Mark end)
+        public DocumentStart(in Mark start, in Mark end)
             : base(start, end)
         {
         }