--- a/YamlDotNet/Core/Tokens/FlowMappingEnd.cs
+++ b/YamlDotNet/Core/Tokens/FlowMappingEnd.cs
@@ -39,7 +39,7 @@ namespace YamlDotNet.Core.Tokens
         /// </summary>
         /// <param name="start">The start position of the token.</param>
         /// <param name="end">The end position of the token.</param>
-        public FlowMappingEnd(Mark start, Mark end)
+        public FlowMappingEnd(in Mark start, in Mark end)
             : base(start, end)
         {
         }