--- a/YamlDotNet/Core/Tokens/FlowMappingStart.cs
+++ b/YamlDotNet/Core/Tokens/FlowMappingStart.cs
@@ -39,7 +39,7 @@ namespace YamlDotNet.Core.Tokens
         /// </summary>
         /// <param name="start">The start position of the token.</param>
         /// <param name="end">The end position of the token.</param>
-        public FlowMappingStart(Mark start, Mark end)
+        public FlowMappingStart(in Mark start, in Mark end)
             : base(start, end)
         {
         }