--- a/YamlDotNet/Core/Tokens/FlowSequenceEnd.cs
+++ b/YamlDotNet/Core/Tokens/FlowSequenceEnd.cs
@@ -39,7 +39,7 @@ namespace YamlDotNet.Core.Tokens
         /// </summary>
         /// <param name="start">The start position of the token.</param>
         /// <param name="end">The end position of the token.</param>
-        public FlowSequenceEnd(Mark start, Mark end)
+        public FlowSequenceEnd(in Mark start, in Mark end)
             : base(start, end)
         {
         }