--- a/YamlDotNet/Core/Tokens/FlowSequenceStart.cs
+++ b/YamlDotNet/Core/Tokens/FlowSequenceStart.cs
@@ -39,7 +39,7 @@ namespace YamlDotNet.Core.Tokens
         /// </summary>
         /// <param name="start">The start position of the token.</param>
         /// <param name="end">The end position of the token.</param>
-        public FlowSequenceStart(Mark start, Mark end)
+        public FlowSequenceStart(in Mark start, in Mark end)
             : base(start, end)
         {
         }