--- a/YamlDotNet/Core/Tokens/Key.cs
+++ b/YamlDotNet/Core/Tokens/Key.cs
@@ -39,7 +39,7 @@ namespace YamlDotNet.Core.Tokens
         /// </summary>
         /// <param name="start">The start position of the token.</param>
         /// <param name="end">The end position of the token.</param>
-        public Key(Mark start, Mark end)
+        public Key(in Mark start, in Mark end)
             : base(start, end)
         {
         }