--- a/YamlDotNet/Core/Tokens/StreamEnd.cs
+++ b/YamlDotNet/Core/Tokens/StreamEnd.cs
@@ -39,7 +39,7 @@ namespace YamlDotNet.Core.Tokens
         /// </summary>
         /// <param name="start">The start position of the token.</param>
         /// <param name="end">The end position of the token.</param>
-        public StreamEnd(Mark start, Mark end)
+        public StreamEnd(in Mark start, in Mark end)
             : base(start, end)
         {
         }