--- a/YamlDotNet/Core/Tokens/Value.cs
+++ b/YamlDotNet/Core/Tokens/Value.cs
@@ -39,7 +39,7 @@ namespace YamlDotNet.Core.Tokens
         /// </summary>
         /// <param name="start">The start position of the token.</param>
         /// <param name="end">The end position of the token.</param>
-        public Value(Mark start, Mark end)
+        public Value(in Mark start, in Mark end)
             : base(start, end)
         {
         }