**!! This is binary breaking change !!**

To reduce memory allocations and improve scanning speed (less GC) I've changed `Mark` to be a read-only struct. In parsing test code profiling showed that over 20% of memory allocations were from constructing `Mark` instances. This is a binary breaking change, but I'd argue the performance benefits are quite good. The API itself didn't change

Using `in` modifier passing `Mark` instances around as struct size exceeds pointer size.

Also introduced `ThrowHelper` to allow inlining `Mark` constructor.

## Benchmark Code

Uses the large YAML from issue https://github.com/aaubry/YamlDotNet/issues/519 . I noticed you don't have a benchmark project currently so I'm running my local one.

```csharp
[MemoryDiagnoser]
public class YamlStreamBenchmark
{
    private string yamlString = "";

    [GlobalSetup]
    public void Setup()
    {
        yamlString = File.ReadAllText(@"c:\work\saltern.yml");
    }

    [Benchmark]
    public void Load()
    {
        var yamlStream = new YamlStream();
        yamlStream.Load(new StringReader(yamlString));
    }
}
```

## Results

``` ini
BenchmarkDotNet=v0.13.1, OS=Windows 10.0.22000
AMD Ryzen 9 5950X, 1 CPU, 32 logical and 16 physical cores
.NET SDK=6.0.100
  [Host]     : .NET 6.0.0 (6.0.21.52210), X64 RyuJIT
  DefaultJob : .NET 6.0.0 (6.0.21.52210), X64 RyuJIT
```

### YamlStreamBenchmark

| **Diff**|Method|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |Load|96.71 ms|1.551 ms|5833.3333|2833.3333|1000.0000|81 MB|
| **New** |	| **73.13 ms (-24%)** | **1.319 ms** | **4428.5714 (-24%)** | **2000.0000 (-29%)** | **714.2857 (-29%)** | **63 MB (-22%)** |