@aaubry 2021-11-15
> I noticed you don't have a benchmark project currently so I'm running my local one.

I used to have one to compare the performance across releases, but it was difficult to maintain and provided little value so it was removed. Feel free to add a benchmark if you feel it would be useful.