@lahma 2021-11-15
@aaubry would you like it as part of this PR or separate?